+++ build.clj
@@ -4,7 +4,7 @@
             [deps-deploy.deps-deploy :as dd]))
 
 (def lib 'net.clojars.idiomattic/easy-onnx)
-(def version (format "0.1.%s" (b/git-count-revs nil)))
+(def version (format "0.2.%s" (b/git-count-revs nil)))
 (def class-dir "target/classes")
 
 (defn test "Run all the tests." [opts]
@@ -19,7 +19,7 @@
   opts)
 
 (defn- pom-template [version]
-  [[:description "A library wrapping com.microsoft.onnxruntime/onnxruntime and ai.djl.huggingface/tokenizers, providing an easy interface for running ML workloads within Clojure applications."]
+  [[:description "A Clojure library for running ONNX models and text embedding via inference4j, plus embedding analysis (DBSCAN clustering, UMAP projection) via smile-core."]
    [:url "https://github.com/idiomattic/easy-onnx"]
    [:licenses
     [:license